move wg init
